add multiple demos on landing page
