+++ usellm.org/app/demo/page.tsx
@@ -1,9 +1,9 @@
-import { HomePageDemo } from "@/components/home-page-demo";
+import { AIChatBotDemo } from "@/components/demos/ai-chatbot-demo";
 
 export default function DemoPage() {
   return (
     <div className="max-w-4xl w-full mx-auto p-4">
-      <HomePageDemo />
+      <AIChatBotDemo />
     </div>
   );
 }
